requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
django-ckeditor-5==0.2.11
django-cors-headers==4.3.1 
django-environ==0.11.2
django-lifecycle==1.1.2
django-mdeditor==0.1.20
django-rest-auth==0.9.5
django-unfold==0.15.1
djangorestframework==3.14.0
django-admin-sortable2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
eskiz-sms
google-auth==2.28.1
google-auth-oauthlib==1.2.0
googletrans==3.1.0a0
psycopg2-binary==2.9.9